Dashboard of Yellow Taxi Trip Records 2019
This dashboard displays some insights from the Yellow Taxi Trip Records data in the last quarter of 2019.  We can see the number of record counts, total amount, average distance, 3 payment types that are most often used, the percentage of vendor id and the comparison of the total amount between the two vendors. There is also a table showing the pickup location id along with the total passengers and distance traveled. From this dashboard we can also see the total extra from October to December 2019.